Add title of shortcuts for Linux .desktop files
New argument -s/--title. When used, will set the parameter `Name` in the
output desktop file. The reason in that the title of the shortcut is
often more verbose than the name of the .desktop file.

If no title is set. The name of the shortcut will be used as the default
value.